refactor: remove barrel import
