--- packages/vitest/src/runtime/runner/index.ts
@@ -1,23 +1,0 @@
-export { recordArtifact } from './artifact'
-export { TestSyntaxError } from './errors'
-export {
-  afterAll,
-  afterEach,
-  aroundAll,
-  aroundEach,
-  beforeAll,
-  beforeEach,
-  onTestFailed,
-  onTestFinished,
-} from './hooks'
-export { getFn, getHooks, setFn, setHooks } from './map'
-export { collectTests, startTests, updateTask } from './run'
-export {
-  createTaskCollector,
-  describe,
-  getCurrentSuite,
-  it,
-  suite,
-  test,
-} from './suite'
-export { getCurrentTest } from './test-state'
